refactor(core): improve gesture tree-shaking and fix review issues
Split gesture.ts into types (gesture.ts) and factory
(create-tap-gesture.ts) so TapRecognizer only bundles when
createTapGesture/createDoubleTapGesture are imported.

- Collapse TapGestureOptions/DoubleTapGestureOptions into GestureOptions
- Fix stale rect in deferred tap/doubletap callbacks (read fresh rect)
- Remove `as EventListener` cast (listen overload infers PointerEvent)
- Remove resolveRegion from barrel export (internal to gesture system)
- Fix stale "coordinator" JSDoc in tap.ts
- Document partial two-region overlap in region.ts
- Add tests: disabled doubletap, first-wins, re-register after cleanup
